backend/src/requirements.txt
wheel
-e tshared_src
-e base3src
psycopg2-binary==2.9.3
tortoise-orm==0.19.0
aerich==0.6.3
aiosmtplib==1.1.6
python-dotenv==0.19.2
tornado==6.1
pycurl==7.45.1 --global-option="--with-openssl"
bcrypt==3.2.0
nanoid==2.0.0
PyJWT==1.7.1
pyjwt-rsa==0.4.0
PyYAML==6.0
asyncpg==0.25.0
python-dateutil==2.8.2
captcha==0.3
pydantic==1.9.0
python-socketio==5.5.2
xmltodict==0.12.0
psycopg2-binary==2.9.3 
coverage
pytest
pandas==1.4.2
openpyxl==3.0.9
Jinja2==3.1.2
click==8.1.3
redis==4.3.3
aiohttp==3.8.1
deepdiff==5.8.1
weasyprint==55.0
PyPDF2==2.4.1
pdf2image==1.16.0
zeep==4.1.0
httpx==0.23.0
beautifulsoup4==4.11.1
asyncclick==8.1.3.2
faker
aioredis==2.0.1
timeago==1.0.16
qrcode==7.3.1
httpx==0.23.0
